% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag.R
\name{lg_id}
\alias{lg_id}
\title{Retrieve lineage IDs from a tagged dataset}
\usage{
lg_id(data)
}
\arguments{
\item{data}{An \code{lg_df} from \code{\link[=lg_tag]{lg_tag()}}.}
}
\value{
A character vector of lineage IDs, one per row.
}
\description{
Returns the \code{lineage_id} vector from an \code{lg_df} object. Use this instead
of accessing the column directly to keep code robust against future
internal changes.
}
\examples{
lg_start()
dm <- data.frame(USUBJID = c("01-001", "01-002"), AGE = c(34L, 52L))
dm_tagged <- lg_tag(dm, dataset_id = "DM")
lg_id(dm_tagged)

}
