% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{lg_operations}
\alias{lg_operations}
\title{Retrieve the operation log as a data frame}
\usage{
lg_operations(verbose = FALSE)
}
\arguments{
\item{verbose}{Logical. Print count summary. Default \code{FALSE}.}
}
\value{
A \code{data.frame} of all recorded operations.
}
\description{
Retrieve the operation log as a data frame
}
